+++ docs/CHANGELOG.rst
@@ -25,6 +25,7 @@ Fixed
 - Set correct executor in flow manager
 - Make executors more robust to unhandled failures
 - Calculate ``Data.size`` by summing ``total_size`` of all file-type outputs
+- Don't change slug explicitly defined by user - raise an error instead
 
 
 ==================